requirements.txt
flask>=2.3.0
flask-socketio>=5.3.0
yfinance>=0.2.31
requests>=2.31.0
pandas>=2.0.0
python-dotenv>=1.0.0
google-genai>=0.3.0
deep-translator>=1.11.0
scikit-learn>=1.3.0
numpy>=1.24.0
